FIX: Prevent infinite update when opening spot editor because callback function for timepciker is created by every render
